Simplify approach, improve views and scripts
