+++ BTCPayServer/Components/StoreLightningBalance/StoreLightningBalance.cs
@@ -24,6 +24,8 @@ public class StoreLightningBalance : ViewComponent
     private readonly CurrencyNameTable _currencies;
     private readonly BTCPayServerOptions _btcpayServerOptions;
     private readonly BTCPayNetworkProvider _networkProvider;
+    private readonly LightningClientFactoryService _lightningClientFactory;
+    private readonly IOptions<LightningNetworkOptions> _lightningNetworkOptions;
     private readonly IOptions<ExternalServicesOptions> _externalServiceOptions;
 
     public StoreLightningBalance(
@@ -40,21 +42,66 @@ public class StoreLightningBalance : ViewComponent
         _networkProvider = networkProvider;
         _btcpayServerOptions = btcpayServerOptions;
         _externalServiceOptions = externalServiceOptions;
+        _lightningClientFactory = lightningClientFactory;
+        _lightningNetworkOptions = lightningNetworkOptions;
         _cryptoCode = _networkProvider.DefaultNetwork.CryptoCode;
     }
 
-    public IViewComponentResult Invoke(StoreData store)
+    public async Task<IViewComponentResult> InvokeAsync(StoreLightningBalanceViewModel vm)
     {
-        var defaultCurrency = store.GetStoreBlob().DefaultCurrency;
+        if (vm.Store == null) throw new ArgumentNullException(nameof(vm.Store));
+        
+        vm.CryptoCode = _cryptoCode;
+        vm.DefaultCurrency = vm.Store.GetStoreBlob().DefaultCurrency;
+        vm.CurrencyData = _currencies.GetCurrencyData(vm.DefaultCurrency, true);
 
-        var vm = new StoreLightningBalanceViewModel
+        if (vm.InitialRendering) return View(vm);
+        
+        try
         {
-            Store = store,
-            CryptoCode = _cryptoCode,
-            CurrencyData = _currencies.GetCurrencyData(defaultCurrency, true),
-            DefaultCurrency = defaultCurrency,
-        };
-
+            var lightningClient = GetLightningClient(vm.Store);
+            var balance = await lightningClient.GetBalance();
+            vm.Balance = balance;
+            vm.TotalOnchain = balance.OnchainBalance != null
+                ? (balance.OnchainBalance.Confirmed?? 0) + (balance.OnchainBalance.Reserved ?? 0) +
+                  (balance.OnchainBalance.Unconfirmed ?? 0)
+                : null;
+            vm.TotalOffchain = balance.OffchainBalance != null
+                ? (balance.OffchainBalance.Opening?? 0) + (balance.OffchainBalance.Local?? 0) +
+                  (balance.OffchainBalance.Closing?? 0)
+                : null;
+        }
+        catch (NotSupportedException)
+        {
+            // not all implementations support balance fetching
+            vm.ProblemDescription = "Your node does not support balance fetching.";
+        }
+        catch
+        {
+            // general error
+            vm.ProblemDescription = "Could not fetch Lightning balance.";
+        }
         return View(vm);
+    }
+    
+    private ILightningClient GetLightningClient(StoreData store)
+    {
+        var network = _networkProvider.GetNetwork<BTCPayNetwork>(_cryptoCode);
+        var id = new PaymentMethodId(_cryptoCode, PaymentTypes.LightningLike);
+        var existing = store.GetSupportedPaymentMethods(_networkProvider)
+            .OfType<LightningSupportedPaymentMethod>()
+            .FirstOrDefault(d => d.PaymentId == id);
+        if (existing == null) return null;
+        
+        if (existing.GetExternalLightningUrl() is {} connectionString)
+        {
+            return _lightningClientFactory.Create(connectionString, network);
+        }
+        if (existing.IsInternalNode && _lightningNetworkOptions.Value.InternalLightningByCryptoCode.TryGetValue(_cryptoCode, out var internalLightningNode))
+        {
+            return _lightningClientFactory.Create(internalLightningNode, network);
+        }
+
+        return null;
     }
 }
