+++ BTCPayServer/Controllers/UIStoresController.cs
@@ -60,8 +60,6 @@ namespace BTCPayServer.Controllers
             AppService appService,
             WebhookSender webhookNotificationManager,
             IDataProtectionProvider dataProtector,
-            LightningClientFactoryService lightningClientFactory,
-            IOptions<LightningNetworkOptions> lightningNetworkOptions,
             IOptions<ExternalServicesOptions> externalServiceOptions)
         {
             _RateFactory = rateFactory;
@@ -84,8 +82,6 @@ namespace BTCPayServer.Controllers
             _BtcpayServerOptions = btcpayServerOptions;
             _BTCPayEnv = btcpayEnv;
             _externalServiceOptions = externalServiceOptions;
-            _lightningNetworkOptions = lightningNetworkOptions;
-            _lightningClientFactory = lightningClientFactory;
         }
 
         readonly BTCPayServerOptions _BtcpayServerOptions;
@@ -106,8 +102,6 @@ namespace BTCPayServer.Controllers
         private readonly AppService _appService;
         private readonly EventAggregator _EventAggregator;
         private readonly IOptions<ExternalServicesOptions> _externalServiceOptions;
-        private readonly IOptions<LightningNetworkOptions> _lightningNetworkOptions;
-        private readonly LightningClientFactoryService _lightningClientFactory;
 
         [TempData]
         public bool StoreNotConfigured
@@ -137,44 +131,6 @@ namespace BTCPayServer.Controllers
         }
 
         public StoreData CurrentStore => HttpContext.GetStoreData();
-
-        [HttpGet("{storeId}")]
-        public async Task<IActionResult> Dashboard()
-        {
-            var store = CurrentStore;
-            var storeBlob = store.GetStoreBlob();
-
-            AddPaymentMethods(store, storeBlob,
-                out var derivationSchemes, out var lightningNodes);
-
-            var walletEnabled = derivationSchemes.Any(scheme => !string.IsNullOrEmpty(scheme.Value) && scheme.Enabled);
-            var lightningEnabled = lightningNodes.Any(ln => !string.IsNullOrEmpty(ln.Address) && ln.Enabled);
-            var vm = new StoreDashboardViewModel
-            {
-                WalletEnabled = walletEnabled,
-                LightningEnabled = lightningEnabled,
-                StoreId = CurrentStore.Id,
-                StoreName = CurrentStore.StoreName,
-                IsSetUp = walletEnabled || lightningEnabled
-            };
-            
-            // Widget data
-            if (vm.WalletEnabled || vm.LightningEnabled)
-            {
-                var userId = GetUserId();
-                var apps = await _appService.GetAllApps(userId, false, store.Id);
-                vm.Apps = apps
-                    .Select(a =>
-                    {
-                        var appData = _appService.GetAppDataIfOwner(userId, a.Id).Result;
-                        appData.StoreData = store;
-                        return appData;
-                    })
-                    .ToList();
-            }
-            
-            return View(vm);
-        }
 
         [HttpPost]
         [Route("{storeId}/users")]
